+++ plone/recipe/codeanalysis/__init__.py
@@ -265,32 +265,38 @@ def code_analysis_flake8(options):
         print("               [\033[00;32m OK \033[0m]")
 
 
+def jshint_errors(output):
+    """JSHint always return an exit code of 2 either if a file has errors or
+    warnings. This method search for markers of errors (E000).
+    """
+    error = re.compile(r'(E\d\d\d)')
+    return error.search(output)
+
+
 def code_analysis_jshint(options):
-    sys.stdout.write("JS Hint")
+    sys.stdout.write("JSHint")
     sys.stdout.flush()
-    files = _find_files(options, '.*\.js')
-    if not files:
-        print("                [\033[00;32m OK \033[0m]")
-        return
 
     # cmd is a sequence of program arguments
     # first argument is child program
     cmd = [
         options['jshint-bin'],
-        '--exclude={0}'.format(options['jshint-exclude']),
+        '--verbose',
+        '--exclude={0}'.format(options['jshint-exclude'] or ' '),
+        options['directory'],
     ]
-    cmd.extend(files.split())
     process = subprocess.Popen(
         cmd,
         stderr=subprocess.STDOUT,
         stdout=subprocess.PIPE
     )
     output, err = process.communicate()
-    if process.returncode:
+    if jshint_errors(output):  # HACK: workaround for JSHint limitations
         print("           [\033[00;31m FAILURE \033[0m]")
         print(output)
     else:
         print("                [\033[00;32m OK \033[0m]")
+        print(output)  # XXX: there should be warnings on the output
 
 
 def code_analysis_csslint(options):