+++ CHANGES.rst
@@ -1,6 +1,10 @@
 1.0a2 (unreleased)
 ------------------
 
+- Workaround over JSHint limitations to avoid displaying warning messages as
+  errors (closes #13).
+  [hvelarde]
+
 - Fix CSS Lint validation and implement new 'csslint-quiet' option.
   [hvelarde]